+++ glyph-brush-layout/CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 * Re-export `rusttype::point`.
 * Fix `bounds_rect` implementation for some `f32::INFINITY` cases.
+* Handle zero & negative scale cases.
 
 # 0.1
 * Initial release.